requirements.txt
dj-database-url==0.5.0
Django==2.2
gunicorn==20.0.4
mysqlclient==1.4.6
Pillow==7.1.2
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.1.0
